refactor: mock 데이터와 핸들러 분리 및 리팩토링
