--- frontend/src/mocks/data.ts
@@ -1,1209 +1,0 @@
-export const myUserData: Member = {
-  id: 1,
-  gitHubId: 'hamcheeseburger',
-  name: '유현지',
-  imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-  careerLevel: 'senior',
-  jobType: 'backend',
-  followerCount: 30,
-  following: false,
-};
-
-export const otherUserData: Member = {
-  id: 1,
-  gitHubId: 'yangdongjue5510',
-  name: 'Dongjue Yang',
-  imageUrl: 'https://avatars.githubusercontent.com/u/87690744?v=4',
-  careerLevel: 'junior',
-  jobType: 'backend',
-  followerCount: 20,
-  following: false,
-};
-
-export const members: ProfileSearchResult[] = [
-  {
-    id: 20,
-    gitHubId: 'deuk960214',
-    name: null,
-    imageUrl: 'https://avatars.githubusercontent.com/u/110340823?v=4',
-    careerLevel: 'none',
-    jobType: 'frontend',
-    followerCount: 20,
-    following: true,
-    profileProducts: [],
-  },
-  {
-    id: 18,
-    gitHubId: 'uk960214',
-    name: 'Yo Wook Kim',
-    imageUrl: 'https://avatars.githubusercontent.com/u/43166681?v=4',
-    careerLevel: 'none',
-    jobType: 'frontend',
-    followerCount: 20,
-    following: false,
-    profileProducts: [],
-  },
-  {
-    id: 1,
-    gitHubId: 'yangdongjue5510',
-    name: 'Dongjue Yang',
-    imageUrl: 'https://avatars.githubusercontent.com/u/87690744?v=4',
-    careerLevel: 'junior',
-    jobType: 'backend',
-    followerCount: 20,
-    following: true,
-    profileProducts: [
-      {
-        id: 12,
-        name: 'K380 멀티 블루투스 키보드',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/959/427/img/3427959_1.jpg?shrink=500:500&_v=20220620102123',
-        reviewCount: 5,
-        rating: 4.2,
-        category: 'keyboard',
-      },
-      {
-        id: 74,
-        name: '마이크로소프트 무선 블루투스 모바일 마우스 3600',
-        imageUrl: 'https://image.auction.co.kr/itemimage/22/98/15/229815dc00.jpg',
-        reviewCount: 1,
-        rating: 4.0,
-        category: 'mouse',
-      },
-      {
-        id: 142,
-        name: '굿밸류 2단 노트북 거치대',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/353/767/img/16767353_1.jpg?shrink=330:330&_v=20220406145257',
-        reviewCount: 3,
-        rating: 5.0,
-        category: 'stand',
-      },
-    ],
-  },
-  {
-    id: 5,
-    gitHubId: 'Ohzzi',
-    name: 'Jihoon Oh',
-    imageUrl: 'https://avatars.githubusercontent.com/u/66253212?v=4',
-    careerLevel: 'none',
-    jobType: 'backend',
-    followerCount: 20,
-    following: true,
-    profileProducts: [
-      {
-        id: 179,
-        name: '요이치 랩킷B 노트북 거치대',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/753/633/img/14633753_1.jpg?shrink=330:330&_v=20210701125159',
-        reviewCount: 1,
-        rating: 5.0,
-        category: 'stand',
-      },
-    ],
-  },
-  {
-    id: 4,
-    gitHubId: 'yh20studio',
-    name: 'tiki',
-    imageUrl: 'https://avatars.githubusercontent.com/u/78459713?v=4',
-    careerLevel: 'none',
-    jobType: 'backend',
-    followerCount: 20,
-    following: false,
-    profileProducts: [],
-  },
-  {
-    id: 2,
-    gitHubId: 'jswith',
-    name: 'Youngmin Hong',
-    imageUrl: 'https://avatars.githubusercontent.com/u/64275588?v=4',
-    careerLevel: 'junior',
-    jobType: 'frontend',
-    followerCount: 20,
-    following: false,
-    profileProducts: [
-      {
-        id: 142,
-        name: '굿밸류 2단 노트북 거치대',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/353/767/img/16767353_1.jpg?shrink=330:330&_v=20220406145257',
-        reviewCount: 3,
-        rating: 5.0,
-        category: 'stand',
-      },
-      {
-        id: 10,
-        name: 'MX KEYS MINI FOR MAC',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/353/332/img/16332353_1.jpg?shrink=500:500&_v=20220207142651',
-        reviewCount: 2,
-        rating: 5.0,
-        category: 'keyboard',
-      },
-      {
-        id: 80,
-        name: 'LG전자 27UP850',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/805/732/img/13732805_1.jpg?shrink=500:500',
-        reviewCount: 1,
-        rating: 5.0,
-        category: 'monitor',
-      },
-      {
-        id: 53,
-        name: 'APPLE Magic Trackpad 2',
-        imageUrl:
-          'https://img.danawa.com/prod_img/500000/983/782/img/16782983_1.jpg?shrink=330:330&_v=20220704092644',
-        reviewCount: 1,
-        rating: 5.0,
-        category: 'mouse',
-      },
-    ],
-  },
-];
-
-export const products: Product[] = [
-  {
-    id: 3,
-    name: '레오폴드 FC900RBT PD 그레이 블루 한글 (저소음 적축)',
-    imageUrl: 'https://upload.wikimedia.org/wikipedia/commons/9/95/Homebrew_logo.svg',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 4,
-    name: '레오폴드 FC900RBT PD 라이트 핑크 한글 (저소음 적축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 5,
-    name: '레오폴드 FC900RBT PD 그라파이트 화이트 한글 (저소음 적축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 6,
-    name: '레오폴드 FC900RBT PD 화이트민트 한글 (저소음적축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 7,
-    name: '레오폴드 FC660MBT PD 라이트 핑크 영문 (저소음 적축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 8,
-    name: '레오폴드 FC750RBT PD 화이트 투톤 한글 (갈축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/631/093/img/16093631_1.jpg?shrink=500:500&_v=20220323142356',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 9,
-    name: '레오폴드 FC900RBT PD 스웨디시 화이트 한글 (청축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/700/573/img/16573700_3.jpg?shrink=500:500&_v=20220311152649',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 10,
-    name: 'MX MECHANICAL',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 11,
-    name: 'SIGNATURE K855',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/signature-k855/gallery/signature-k855-graphite-kor-gallery-1.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 12,
-    name: 'MX KEYS MINI FOR MAC',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-keys-mini-for-mac/gallery/kor/mx-keys-mini-top-mac-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 13,
-    name: 'MX KEYS MINI',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 14,
-    name: 'K380 멀티 블루투스 키보드',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 15,
-    name: 'MX KEYS 키보드',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 16,
-    name: 'ERGO K860',
-    imageUrl:
-      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 17,
-    name: '앱코 HACKER K660 축교환 완전방수 게이밍 카일 광축 블랙 (클릭)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/479/221/img/5221479_1.jpg?shrink=500:500&_v=20220216130109',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 18,
-    name: '앱코 HACKER K640 축교환 게이밍 기계식 화이트 (청축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/947/128/img/5128947_1.jpg?shrink=500:500&_v=20210930103632',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 19,
-    name: '앱코 HACKER K8700 RED LED 에디션 카일 광축 축교환 완전방수 게이밍',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/466/771/img/6771466_1.jpg?shrink=500:500&_v=20190805150045',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 20,
-    name: '앱코 KN10C PBT 무접점 키보드',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/799/229/img/15229799_1.jpg?shrink=500:500&_v=20210913112802',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 21,
-    name: '앱코 HACKER K300 게이밍 키보드',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/321/248/img/3248321_1.jpg?shrink=500:500&_v=20190806133811',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 22,
-    name: '앱코 KN10 BT 블루투스 무접점 게이밍 키보드 화이트',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/574/376/img/12376574_1.jpg?shrink=500:500&_v=20210930110302',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 23,
-    name: '앱코 HACKER K990 V3 스카이문 LED 완전방수 무접점',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/871/242/img/9242871_1.jpg?shrink=500:500&_v=20200107105606',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 24,
-    name: 'COX 게이트론 LED 게이밍 기계식키보드 황축',
-    imageUrl:
-      'https://encrypted-tbn2.gstatic.com/shopping?q=tbn:ANd9GcQiEvjOVYbzG3CDnczx9tKYwV47j9HGwFaz1pReYIypkyUnxUoJKdOvuotL0JvkJkc1UMcbHhrABw&usqp=CAc',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 25,
-    name: 'COX 축교환 기계식 게이밍 키보드 저소음백축',
-    imageUrl:
-      'https://encrypted-tbn2.gstatic.com/shopping?q=tbn:ANd9GcQwUgGiKjSCM19UGv0oAuT2R8AtaTTmVfEV-MsxVj_8I-EuPn4JiXYhAadKjzBJHLjGWU6TKZjFfg&usqp=CAc',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 26,
-    name: '콕스 게이트론 LED 게이밍 기계식 키보드 황축',
-    imageUrl:
-      'https://encrypted-tbn3.gstatic.com/shopping?q=tbn:ANd9GcTXb4d46RawSgTA9pWb6jvPR9s77elV8XKgwvsVb6n1MLWp57yR_MNrvWkTt9HUlnCWyHxbg9QQ0A&usqp=CAc',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 27,
-    name: 'COX CK87 레드/그레이 게이트론 LED 게이밍 기계식',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/979/938/img/7938979_1.jpg?shrink=130:130&_v=20190916165446',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 28,
-    name: 'COX CK01 TKL',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/445/440/img/14440445_1.jpg?shrink=130:130&_v=20210610172313',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 29,
-    name: 'COX CY104 게이트론 기계식 키보드',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/237/828/img/14828237_1.jpg?shrink=130:130&_v=20220414111601',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 30,
-    name: 'COX CA106 CNC 알루미늄 어고노믹 게이트론 기계식 키보드',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/557/640/img/11640557_1.jpg?shrink=130:130&_v=20200625090716',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 31,
-    name: '한성컴퓨터 GK200 RAINBOW 게이밍 기계식 키보드 (블랙, 청축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/064/357/img/16357064_1.jpg?shrink=330:330&_v=20220602181027',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 32,
-    name: '한성컴퓨터 GK787S OfficeMaster 체리 기계식 키보드 (블랙, 저소음 적축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 33,
-    name: '한성컴퓨터 BK330 멀티페어링 블루투스 키보드 (블랙)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/640/413/img/14413640_1.jpg?shrink=330:330&_v=20220502170755',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 34,
-    name: '한성컴퓨터 GK898B OfficeMaster Gray Edition',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/676/071/img/17071676_1.jpg?shrink=330:330&_v=20220518123903',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 35,
-    name: '한성컴퓨터 GK687 TKL 게이트론 RGB 기계식 키보드 (화이트, 청축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/106/908/img/14908106_1.jpg?shrink=330:330&_v=20210803164239',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 36,
-    name: '한성컴퓨터 GTune MKF30S RAINBOW (블랙, 청축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/224/291/img/5291224_1.jpg?shrink=330:330&_v=20210520103656',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 37,
-    name: '한성컴퓨터 GTune MBF77 Vision (블랙)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/018/011/img/5011018_1.jpg?shrink=330:330&_v=20211020171107',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-  {
-    id: 38,
-    name: '한성컴퓨터 GK998B SKY 블루투스 기계식 키보드 영문 (화이트, 갈축)',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/440/246/img/12246440_1.jpg?shrink=330:330&_v=20210813173050',
-    rating: 0.0,
-    reviewCount: 3,
-    category: 'keyboard',
-  },
-];
-
-const getReviewProductData: (
-  product: Product
-) => Pick<Product, 'id' | 'name' | 'imageUrl'> = ({ id, name, imageUrl }) => ({
-  id,
-  name,
-  imageUrl,
-});
-
-export const InventoryProducts: { items: InventoryProduct[] } = {
-  items: [
-    {
-      id: 1,
-      selected: true,
-      product: {
-        id: 1,
-        name: '키보드1',
-        imageUrl:
-          'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e9/Jenkins_logo.svg/1200px-Jenkins_logo.svg.png',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'keyboard',
-      },
-    },
-    {
-      id: 2,
-      selected: false,
-      product: {
-        id: 2,
-        name: '키보드2',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'keyboard',
-      },
-    },
-    {
-      id: 3,
-      selected: true,
-      product: {
-        id: 3,
-        name: '마우스1',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'mouse',
-      },
-    },
-    {
-      id: 4,
-      selected: false,
-      product: {
-        id: 4,
-        name: '마우스2',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'mouse',
-      },
-    },
-    {
-      id: 5,
-      selected: false,
-      product: {
-        id: 5,
-        name: '모니터1',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'monitor',
-      },
-    },
-    {
-      id: 6,
-      selected: true,
-      product: {
-        id: 6,
-        name: '모니터2',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/833/414/img/15414833_1.jpg?shrink=330:330&_v=20211013110445',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'monitor',
-      },
-    },
-    {
-      id: 7,
-      selected: false,
-      product: {
-        id: 7,
-        name: '거치대1',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'stand',
-      },
-    },
-    {
-      id: 8,
-      selected: true,
-      product: {
-        id: 8,
-        name: '거치대2',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'stand',
-      },
-    },
-    {
-      id: 9,
-      selected: false,
-      product: {
-        id: 9,
-        name: '소프트웨어1',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'software',
-      },
-    },
-    {
-      id: 10,
-      selected: false,
-      product: {
-        id: 10,
-        name: '소프트웨어2',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'software',
-      },
-    },
-    {
-      id: 10,
-      selected: false,
-      product: {
-        id: 10,
-        name: '키보드3',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'keyboard',
-      },
-    },
-    {
-      id: 11,
-      selected: false,
-      product: {
-        id: 11,
-        name: '키보드4',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'keyboard',
-      },
-    },
-    {
-      id: 12,
-      selected: false,
-      product: {
-        id: 12,
-        name: '키보드5',
-        imageUrl:
-          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-        reviewCount: 0,
-        rating: 0.0,
-        category: 'keyboard',
-      },
-    },
-  ],
-};
-
-export const InventoryReview = {
-  id: 1,
-  product: {
-    id: 1,
-    name: '키보드1',
-    imageUrl:
-      'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
-    reviewCount: 2,
-    rating: 2,
-    category: 'keyboard',
-  },
-  content:
-    '리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다리뷰입니다',
-  rating: 4,
-  createdAt: '2022-08-18T14:39:29.686202',
-};
-
-export const reviewsWithProduct: Review[] = [
-  {
-    id: 1,
-    author: {
-      id: 1,
-      gitHubId: '사용자1',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 1,
-      name: '키보드1',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content: `sdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflkvsdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflkvsdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlks
-
-
-
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflk
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflksdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    sdklfjasdfklsfndlkskaldflk
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    `,
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 2,
-    author: {
-      id: 2,
-      gitHubId: '사용자2',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 2,
-      name: '키보드2',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 3,
-    author: {
-      id: 3,
-      gitHubId: '사용자3',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 3,
-      name: '키보드3',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 4,
-    author: {
-      id: 4,
-      gitHubId: '사용자4',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 4,
-      name: '키보드4',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 5,
-    author: {
-      id: 5,
-      gitHubId: '사용자5',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 5,
-      name: '키보드5',
-      imageUrl: 'https://upload.wikimedia.org/wikipedia/commons/9/95/Homebrew_logo.svg',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 6,
-    author: {
-      id: 6,
-      gitHubId: '사용자6',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 6,
-      name: '키보드6',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 7,
-    author: {
-      id: 7,
-      gitHubId: '사용자7',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 7,
-      name: '키보드7',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 8,
-    author: {
-      id: 8,
-      gitHubId: '사용자8',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 8,
-      name: '키보드8',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 9,
-    author: {
-      id: 9,
-      gitHubId: '사용자9',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 9,
-      name: '키보드9',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 10,
-    author: {
-      id: 10,
-      gitHubId: '사용자10',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 10,
-      name: '키보드10',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 11,
-    author: {
-      id: 11,
-      gitHubId: '사용자11',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 11,
-      name: '키보드11',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 12,
-    author: {
-      id: 12,
-      gitHubId: '사용자12',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    product: {
-      id: 12,
-      name: '키보드12',
-      imageUrl:
-        'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
-    },
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-];
-
-export const reviewsWithOutProduct: Review[] = [
-  {
-    id: 1,
-    author: {
-      id: 1,
-      gitHubId: '사용자1',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 1,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: true,
-  },
-  {
-    id: 2,
-    author: {
-      id: 2,
-      gitHubId: '사용자2',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 2,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 3,
-    author: {
-      id: 3,
-      gitHubId: '사용자3',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 3,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 4,
-    author: {
-      id: 4,
-      gitHubId: '사용자4',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 4,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 5,
-    author: {
-      id: 5,
-      gitHubId: '사용자5',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 5,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 6,
-    author: {
-      id: 6,
-      gitHubId: '사용자6',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 6,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 7,
-    author: {
-      id: 7,
-      gitHubId: '사용자7',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 7,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 8,
-    author: {
-      id: 8,
-      gitHubId: '사용자8',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 8,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 9,
-    author: {
-      id: 9,
-      gitHubId: '사용자9',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 9,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 10,
-    author: {
-      id: 10,
-      gitHubId: '사용자10',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 10,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 11,
-    author: {
-      id: 11,
-      gitHubId: '사용자11',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 11,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-  {
-    id: 12,
-    author: {
-      id: 12,
-      gitHubId: '사용자12',
-      imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
-    },
-    productId: 12,
-    content:
-      '무접점은 처음 사용이라 바로 적응되진 않아요 그래도 검증된 제품이라 역시 좋긴 좋네요 작업용으로 마지막 키보드라 생각한거라 비싸도 확 질렀습니다 아는 분은 아시겠지만 제품이 국내로 넘어온 후 관세청에서 문자로 제세액 3만원이상의 금액을 입금하라고 오더라구요 ',
-    rating: 5,
-    createdAt: '2022-07-03 14:23:23',
-    authorMatch: false,
-  },
-];
