+++ frontend/src/mocks/data/inventoryProducts.ts
@@ -1,0 +1,173 @@
+export const InventoryProducts: { items: InventoryProduct[] } = {
+  items: [
+    {
+      id: 1,
+      selected: true,
+      product: {
+        id: 1,
+        name: '키보드1',
+        imageUrl:
+          'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e9/Jenkins_logo.svg/1200px-Jenkins_logo.svg.png',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'keyboard',
+      },
+    },
+    {
+      id: 2,
+      selected: false,
+      product: {
+        id: 2,
+        name: '키보드2',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'keyboard',
+      },
+    },
+    {
+      id: 3,
+      selected: true,
+      product: {
+        id: 3,
+        name: '마우스1',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'mouse',
+      },
+    },
+    {
+      id: 4,
+      selected: false,
+      product: {
+        id: 4,
+        name: '마우스2',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'mouse',
+      },
+    },
+    {
+      id: 5,
+      selected: false,
+      product: {
+        id: 5,
+        name: '모니터1',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'monitor',
+      },
+    },
+    {
+      id: 6,
+      selected: true,
+      product: {
+        id: 6,
+        name: '모니터2',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/833/414/img/15414833_1.jpg?shrink=330:330&_v=20211013110445',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'monitor',
+      },
+    },
+    {
+      id: 7,
+      selected: false,
+      product: {
+        id: 7,
+        name: '거치대1',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'stand',
+      },
+    },
+    {
+      id: 8,
+      selected: true,
+      product: {
+        id: 8,
+        name: '거치대2',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'stand',
+      },
+    },
+    {
+      id: 9,
+      selected: false,
+      product: {
+        id: 9,
+        name: '소프트웨어1',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'software',
+      },
+    },
+    {
+      id: 10,
+      selected: false,
+      product: {
+        id: 10,
+        name: '소프트웨어2',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'software',
+      },
+    },
+    {
+      id: 10,
+      selected: false,
+      product: {
+        id: 10,
+        name: '키보드3',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'keyboard',
+      },
+    },
+    {
+      id: 11,
+      selected: false,
+      product: {
+        id: 11,
+        name: '키보드4',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'keyboard',
+      },
+    },
+    {
+      id: 12,
+      selected: false,
+      product: {
+        id: 12,
+        name: '키보드5',
+        imageUrl:
+          'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+        reviewCount: 0,
+        rating: 0.0,
+        category: 'keyboard',
+      },
+    },
+  ],
+};
