+++ frontend/src/mocks/data/member.ts
@@ -1,0 +1,164 @@
+export const myUserData: Member = {
+  id: 1,
+  gitHubId: 'hamcheeseburger',
+  name: '유현지',
+  imageUrl: 'https://avatars.githubusercontent.com/u/61769743?v=4',
+  careerLevel: 'senior',
+  jobType: 'backend',
+  followerCount: 30,
+  following: false,
+};
+
+export const otherUserData: Member = {
+  id: 1,
+  gitHubId: 'yangdongjue5510',
+  name: 'Dongjue Yang',
+  imageUrl: 'https://avatars.githubusercontent.com/u/87690744?v=4',
+  careerLevel: 'junior',
+  jobType: 'backend',
+  followerCount: 20,
+  following: false,
+};
+
+export const members: ProfileSearchResult[] = [
+  {
+    id: 20,
+    gitHubId: 'deuk960214',
+    name: null,
+    imageUrl: 'https://avatars.githubusercontent.com/u/110340823?v=4',
+    careerLevel: 'none',
+    jobType: 'frontend',
+    followerCount: 20,
+    following: true,
+    profileProducts: [],
+  },
+  {
+    id: 18,
+    gitHubId: 'uk960214',
+    name: 'Yo Wook Kim',
+    imageUrl: 'https://avatars.githubusercontent.com/u/43166681?v=4',
+    careerLevel: 'none',
+    jobType: 'frontend',
+    followerCount: 20,
+    following: false,
+    profileProducts: [],
+  },
+  {
+    id: 1,
+    gitHubId: 'yangdongjue5510',
+    name: 'Dongjue Yang',
+    imageUrl: 'https://avatars.githubusercontent.com/u/87690744?v=4',
+    careerLevel: 'junior',
+    jobType: 'backend',
+    followerCount: 20,
+    following: true,
+    profileProducts: [
+      {
+        id: 12,
+        name: 'K380 멀티 블루투스 키보드',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/959/427/img/3427959_1.jpg?shrink=500:500&_v=20220620102123',
+        reviewCount: 5,
+        rating: 4.2,
+        category: 'keyboard',
+      },
+      {
+        id: 74,
+        name: '마이크로소프트 무선 블루투스 모바일 마우스 3600',
+        imageUrl: 'https://image.auction.co.kr/itemimage/22/98/15/229815dc00.jpg',
+        reviewCount: 1,
+        rating: 4.0,
+        category: 'mouse',
+      },
+      {
+        id: 142,
+        name: '굿밸류 2단 노트북 거치대',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/353/767/img/16767353_1.jpg?shrink=330:330&_v=20220406145257',
+        reviewCount: 3,
+        rating: 5.0,
+        category: 'stand',
+      },
+    ],
+  },
+  {
+    id: 5,
+    gitHubId: 'Ohzzi',
+    name: 'Jihoon Oh',
+    imageUrl: 'https://avatars.githubusercontent.com/u/66253212?v=4',
+    careerLevel: 'none',
+    jobType: 'backend',
+    followerCount: 20,
+    following: true,
+    profileProducts: [
+      {
+        id: 179,
+        name: '요이치 랩킷B 노트북 거치대',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/753/633/img/14633753_1.jpg?shrink=330:330&_v=20210701125159',
+        reviewCount: 1,
+        rating: 5.0,
+        category: 'stand',
+      },
+    ],
+  },
+  {
+    id: 4,
+    gitHubId: 'yh20studio',
+    name: 'tiki',
+    imageUrl: 'https://avatars.githubusercontent.com/u/78459713?v=4',
+    careerLevel: 'none',
+    jobType: 'backend',
+    followerCount: 20,
+    following: false,
+    profileProducts: [],
+  },
+  {
+    id: 2,
+    gitHubId: 'jswith',
+    name: 'Youngmin Hong',
+    imageUrl: 'https://avatars.githubusercontent.com/u/64275588?v=4',
+    careerLevel: 'junior',
+    jobType: 'frontend',
+    followerCount: 20,
+    following: false,
+    profileProducts: [
+      {
+        id: 142,
+        name: '굿밸류 2단 노트북 거치대',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/353/767/img/16767353_1.jpg?shrink=330:330&_v=20220406145257',
+        reviewCount: 3,
+        rating: 5.0,
+        category: 'stand',
+      },
+      {
+        id: 10,
+        name: 'MX KEYS MINI FOR MAC',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/353/332/img/16332353_1.jpg?shrink=500:500&_v=20220207142651',
+        reviewCount: 2,
+        rating: 5.0,
+        category: 'keyboard',
+      },
+      {
+        id: 80,
+        name: 'LG전자 27UP850',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/805/732/img/13732805_1.jpg?shrink=500:500',
+        reviewCount: 1,
+        rating: 5.0,
+        category: 'monitor',
+      },
+      {
+        id: 53,
+        name: 'APPLE Magic Trackpad 2',
+        imageUrl:
+          'https://img.danawa.com/prod_img/500000/983/782/img/16782983_1.jpg?shrink=330:330&_v=20220704092644',
+        reviewCount: 1,
+        rating: 5.0,
+        category: 'mouse',
+      },
+    ],
+  },
+];
