+++ frontend/src/mocks/data/products.ts
@@ -1,0 +1,325 @@
+export const products: Product[] = [
+  {
+    id: 3,
+    name: '레오폴드 FC900RBT PD 그레이 블루 한글 (저소음 적축)',
+    imageUrl: 'https://upload.wikimedia.org/wikipedia/commons/9/95/Homebrew_logo.svg',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 4,
+    name: '레오폴드 FC900RBT PD 라이트 핑크 한글 (저소음 적축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/701/350/img/16350701_1.jpg?shrink=330:330&_v=20220209091825',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 5,
+    name: '레오폴드 FC900RBT PD 그라파이트 화이트 한글 (저소음 적축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/637/448/img/15448637_1.jpg?shrink=330:330&_v=20211015164726',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 6,
+    name: '레오폴드 FC900RBT PD 화이트민트 한글 (저소음적축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/701/079/img/17079701_1.jpg?shrink=330:330&_v=20220519094414',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 7,
+    name: '레오폴드 FC660MBT PD 라이트 핑크 영문 (저소음 적축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/578/350/img/16350578_1.jpg?shrink=500:500&_v=20220209090426',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 8,
+    name: '레오폴드 FC750RBT PD 화이트 투톤 한글 (갈축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/631/093/img/16093631_1.jpg?shrink=500:500&_v=20220323142356',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 9,
+    name: '레오폴드 FC900RBT PD 스웨디시 화이트 한글 (청축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/700/573/img/16573700_3.jpg?shrink=500:500&_v=20220311152649',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 10,
+    name: 'MX MECHANICAL',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 11,
+    name: 'SIGNATURE K855',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/signature-k855/gallery/signature-k855-graphite-kor-gallery-1.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 12,
+    name: 'MX KEYS MINI FOR MAC',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-keys-mini-for-mac/gallery/kor/mx-keys-mini-top-mac-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 13,
+    name: 'MX KEYS MINI',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 14,
+    name: 'K380 멀티 블루투스 키보드',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 15,
+    name: 'MX KEYS 키보드',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 16,
+    name: 'ERGO K860',
+    imageUrl:
+      'https://resource.logitech.com/w_386,ar_1.0,c_limit,f_auto,q_auto,dpr_2.0/d_transparent.gif/content/dam/logitech/en/products/keyboards/mx-mechanical/gallery/mx-mechanical-keyboard-top-view-graphite-kor.png?v=1',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 17,
+    name: '앱코 HACKER K660 축교환 완전방수 게이밍 카일 광축 블랙 (클릭)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/479/221/img/5221479_1.jpg?shrink=500:500&_v=20220216130109',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 18,
+    name: '앱코 HACKER K640 축교환 게이밍 기계식 화이트 (청축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/947/128/img/5128947_1.jpg?shrink=500:500&_v=20210930103632',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 19,
+    name: '앱코 HACKER K8700 RED LED 에디션 카일 광축 축교환 완전방수 게이밍',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/466/771/img/6771466_1.jpg?shrink=500:500&_v=20190805150045',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 20,
+    name: '앱코 KN10C PBT 무접점 키보드',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/799/229/img/15229799_1.jpg?shrink=500:500&_v=20210913112802',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 21,
+    name: '앱코 HACKER K300 게이밍 키보드',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/321/248/img/3248321_1.jpg?shrink=500:500&_v=20190806133811',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 22,
+    name: '앱코 KN10 BT 블루투스 무접점 게이밍 키보드 화이트',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/574/376/img/12376574_1.jpg?shrink=500:500&_v=20210930110302',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 23,
+    name: '앱코 HACKER K990 V3 스카이문 LED 완전방수 무접점',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/871/242/img/9242871_1.jpg?shrink=500:500&_v=20200107105606',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 24,
+    name: 'COX 게이트론 LED 게이밍 기계식키보드 황축',
+    imageUrl:
+      'https://encrypted-tbn2.gstatic.com/shopping?q=tbn:ANd9GcQiEvjOVYbzG3CDnczx9tKYwV47j9HGwFaz1pReYIypkyUnxUoJKdOvuotL0JvkJkc1UMcbHhrABw&usqp=CAc',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 25,
+    name: 'COX 축교환 기계식 게이밍 키보드 저소음백축',
+    imageUrl:
+      'https://encrypted-tbn2.gstatic.com/shopping?q=tbn:ANd9GcQwUgGiKjSCM19UGv0oAuT2R8AtaTTmVfEV-MsxVj_8I-EuPn4JiXYhAadKjzBJHLjGWU6TKZjFfg&usqp=CAc',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 26,
+    name: '콕스 게이트론 LED 게이밍 기계식 키보드 황축',
+    imageUrl:
+      'https://encrypted-tbn3.gstatic.com/shopping?q=tbn:ANd9GcTXb4d46RawSgTA9pWb6jvPR9s77elV8XKgwvsVb6n1MLWp57yR_MNrvWkTt9HUlnCWyHxbg9QQ0A&usqp=CAc',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 27,
+    name: 'COX CK87 레드/그레이 게이트론 LED 게이밍 기계식',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/979/938/img/7938979_1.jpg?shrink=130:130&_v=20190916165446',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 28,
+    name: 'COX CK01 TKL',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/445/440/img/14440445_1.jpg?shrink=130:130&_v=20210610172313',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 29,
+    name: 'COX CY104 게이트론 기계식 키보드',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/237/828/img/14828237_1.jpg?shrink=130:130&_v=20220414111601',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 30,
+    name: 'COX CA106 CNC 알루미늄 어고노믹 게이트론 기계식 키보드',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/557/640/img/11640557_1.jpg?shrink=130:130&_v=20200625090716',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 31,
+    name: '한성컴퓨터 GK200 RAINBOW 게이밍 기계식 키보드 (블랙, 청축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/064/357/img/16357064_1.jpg?shrink=330:330&_v=20220602181027',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 32,
+    name: '한성컴퓨터 GK787S OfficeMaster 체리 기계식 키보드 (블랙, 저소음 적축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/784/731/img/11731784_1.jpg?shrink=330:330&_v=20220602181027',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 33,
+    name: '한성컴퓨터 BK330 멀티페어링 블루투스 키보드 (블랙)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/640/413/img/14413640_1.jpg?shrink=330:330&_v=20220502170755',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 34,
+    name: '한성컴퓨터 GK898B OfficeMaster Gray Edition',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/676/071/img/17071676_1.jpg?shrink=330:330&_v=20220518123903',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 35,
+    name: '한성컴퓨터 GK687 TKL 게이트론 RGB 기계식 키보드 (화이트, 청축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/106/908/img/14908106_1.jpg?shrink=330:330&_v=20210803164239',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 36,
+    name: '한성컴퓨터 GTune MKF30S RAINBOW (블랙, 청축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/224/291/img/5291224_1.jpg?shrink=330:330&_v=20210520103656',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 37,
+    name: '한성컴퓨터 GTune MBF77 Vision (블랙)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/018/011/img/5011018_1.jpg?shrink=330:330&_v=20211020171107',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+  {
+    id: 38,
+    name: '한성컴퓨터 GK998B SKY 블루투스 기계식 키보드 영문 (화이트, 갈축)',
+    imageUrl:
+      'http://img.danawa.com/prod_img/500000/440/246/img/12246440_1.jpg?shrink=330:330&_v=20210813173050',
+    rating: 0.0,
+    reviewCount: 3,
+    category: 'keyboard',
+  },
+];
